+++ src/lpj/celldata.c
@@ -122,7 +122,7 @@ Celldata opencelldata(Config *config /**< LPJmL configuration */
   if(config->with_lakes)
   {
     /* Open file for lake fraction */
-    if(openinputdata(&celldata->lakes,&config->lakes_filename,"lakes","1",LPJ_BYTE,0.01,config))
+    if(openinputdata(&celldata->lakes,&config->lakes_filename,"lakes","1",LPJ_BYTE,0.01,0,config))
     {
       if(config->soil_filename.fmt==CDF)
         closecoord_netcdf(celldata->soil.cdf);
@@ -137,7 +137,7 @@ Celldata opencelldata(Config *config /**< LPJmL configuration */
   }
   if(config->with_nitrogen)
   {
-    if(openinputdata(&celldata->soilph,&config->soilph_filename,"soilph",NULL,LPJ_SHORT,0.01,config))
+    if(openinputdata(&celldata->soilph,&config->soilph_filename,"soilph",NULL,LPJ_SHORT,0.01,0,config))
     {
       if(config->soil_filename.fmt==CDF)
         closecoord_netcdf(celldata->soil.cdf);
@@ -154,7 +154,7 @@ Celldata opencelldata(Config *config /**< LPJmL configuration */
   }
   if(config->landfrac_from_file)
   {
-    if(openinputdata(&celldata->landfrac,&config->landfrac_filename,"landfrac","1",LPJ_SHORT,0.01,config))
+    if(openinputdata(&celldata->landfrac,&config->landfrac_filename,"landfrac","1",LPJ_SHORT,0.01,0,config))
     {
       if(config->soil_filename.fmt==CDF)
         closecoord_netcdf(celldata->soil.cdf);